chore: use uint256 for token incdices

+++ src/tests/Space.t.sol
@@ -3,22 +3,22 @@ pragma solidity ^0.7.0;
 pragma experimental ABIEncoderV2;
 
 // Testing utils
-import { DSTest } from "@sense-finance/v1-core/src/tests/test-helpers/DSTest.sol";
-import { MockDividerSpace, MockAdapterSpace, ERC20Mintable } from "./utils/Mocks.sol";
-import { VM } from "./utils/VM.sol";
-import { User } from "./utils/User.sol";
+import {DSTest} from "@sense-finance/v1-core/src/tests/test-helpers/DSTest.sol";
+import {MockDividerSpace, MockAdapterSpace, ERC20Mintable} from "./utils/Mocks.sol";
+import {VM} from "./utils/VM.sol";
+import {User} from "./utils/User.sol";
 
 // External references
-import { Vault, IVault, IWETH, IAuthorizer, IAsset, IProtocolFeesCollector } from "@balancer-labs/v2-vault/contracts/Vault.sol";
-import { Authentication } from "@balancer-labs/v2-solidity-utils/contracts/helpers/Authentication.sol";
-import { IERC20 } from "@balancer-labs/v2-solidity-utils/contracts/openzeppelin/IERC20.sol";
-import { Authorizer } from "@balancer-labs/v2-vault/contracts/Authorizer.sol";
-import { FixedPoint } from "@balancer-labs/v2-solidity-utils/contracts/math/FixedPoint.sol";
+import {Vault, IVault, IWETH, IAuthorizer, IAsset, IProtocolFeesCollector} from "@balancer-labs/v2-vault/contracts/Vault.sol";
+import {Authentication} from "@balancer-labs/v2-solidity-utils/contracts/helpers/Authentication.sol";
+import {IERC20} from "@balancer-labs/v2-solidity-utils/contracts/openzeppelin/IERC20.sol";
+import {Authorizer} from "@balancer-labs/v2-vault/contracts/Authorizer.sol";
+import {FixedPoint} from "@balancer-labs/v2-solidity-utils/contracts/math/FixedPoint.sol";
 
 // Internal references
-import { SpaceFactory } from "../SpaceFactory.sol";
-import { Space } from "../Space.sol";
-import { Errors } from "../Errors.sol";
+import {SpaceFactory} from "../SpaceFactory.sol";
+import {Space} from "../Space.sol";
+import {Errors} from "../Errors.sol";
 
 // Base DSTest plus a few extra features
 contract Test is DSTest {
@@ -50,7 +50,8 @@ contract SpaceTest is Test {
     using FixedPoint for uint256;
 
     VM internal constant vm = VM(HEVM_ADDRESS);
-    IWETH internal constant weth = IWETH(0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2);
+    IWETH internal constant weth =
+        IWETH(0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2);
     uint256 public constant INTIAL_USER_BALANCE = 100e18;
 
     Vault internal vault;
@@ -95,7 +96,10 @@ contract SpaceTest is Test {
 
         space = Space(spaceFactory.create(address(adapter), maturity));
 
-        (address _zero, , , , , , , , ) = MockDividerSpace(divider).series(address(adapter), maturity);
+        (address _zero, , , , , , , , ) = MockDividerSpace(divider).series(
+            address(adapter),
+            maturity
+        );
         zero = ERC20Mintable(_zero);
         target = ERC20Mintable(adapter.target());
 
@@ -172,8 +176,10 @@ contract SpaceTest is Test {
         // Received less than one Target
         assertEq(targetOt, expectedTargetOut);
 
-        (, uint256[] memory balances, ) = vault.getPoolTokens(space.getPoolId());
-        (uint8 zeroi, uint8 targeti) = space.getIndices();
+        (, uint256[] memory balances, ) = vault.getPoolTokens(
+            space.getPoolId()
+        );
+        (uint256 zeroi, uint256 targeti) = space.getIndices();
 
         // Pool balances reflect the user's balances
         assertEq(balances[zeroi], 1e18);
@@ -190,7 +196,10 @@ contract SpaceTest is Test {
         uint256 zeroOut = jim.swapIn(false, 0.5e18);
         uint256 expectedZeroOut = 804788983856768174;
 
-        assertEq(target.balanceOf(address(jim)), 99e18 + expectedTargetOut - 0.5e18);
+        assertEq(
+            target.balanceOf(address(jim)),
+            99e18 + expectedTargetOut - 0.5e18
+        );
         assertEq(zeroOut, expectedZeroOut);
     }
 
@@ -270,11 +279,16 @@ contract SpaceTest is Test {
         ava.join();
         assertGe(target.balanceOf(address(ava)), 99e18);
         // Should have joined less Target than last time
-        assertGt(100e18 - targetPreJoin, targetPreJoin - target.balanceOf(address(ava)));
+        assertGt(
+            100e18 - targetPreJoin,
+            targetPreJoin - target.balanceOf(address(ava))
+        );
         // Should have joined Target / Zeros at the ratio of the pool
         assertEq(zero.balanceOf(address(ava)), 98e18);
-        (, uint256[] memory balances, ) = vault.getPoolTokens(space.getPoolId());
-        (uint8 zeroi, uint8 targeti) = space.getIndices();
+        (, uint256[] memory balances, ) = vault.getPoolTokens(
+            space.getPoolId()
+        );
+        (uint256 zeroi, uint256 targeti) = space.getIndices();
         // All tokens are 18 decimals in `setUp`
         uint256 targetPerZero = (balances[targeti] * 1e18) / balances[zeroi];
         // TargetPerZero * 1 = Target amount in for 1 Zero in
@@ -364,7 +378,8 @@ contract SpaceTest is Test {
         }
 
         jim.exit(space.balanceOf(address(jim)));
-        uint256 currentPositionValue = target.balanceOf(address(jim)) + zero.balanceOf(address(jim)).mulDown(zeroPrice);
+        uint256 currentPositionValue = target.balanceOf(address(jim)) +
+            zero.balanceOf(address(jim)).mulDown(zeroPrice);
         assertGt(currentPositionValue, startingPositionValue);
     }
 
@@ -381,7 +396,11 @@ contract SpaceTest is Test {
         vm.warp(maturity - 1);
 
         assertClose(sid.swapIn(true).mulDown(adapter.scale()), 1e18, 1e11);
-        assertClose(sid.swapIn(false, uint256(1e18).divDown(adapter.scale())), 1e18, 1e11);
+        assertClose(
+            sid.swapIn(false, uint256(1e18).divDown(adapter.scale())),
+            1e18,
+            1e11
+        );
     }
 
     function testConstantSumAfterMaturity() public {
@@ -397,7 +416,11 @@ contract SpaceTest is Test {
         vm.warp(maturity + 1);
 
         assertClose(sid.swapIn(true).mulDown(adapter.scale()), 1e18, 1e7);
-        assertClose(sid.swapIn(false, uint256(1e18).divDown(adapter.scale())), 1e18, 1e7);
+        assertClose(
+            sid.swapIn(false, uint256(1e18).divDown(adapter.scale())),
+            1e18,
+            1e7
+        );
     }
 
     function testCantJoinAfterMaturity() public {
@@ -411,12 +434,12 @@ contract SpaceTest is Test {
     }
 
     function testProtocolFees() public {
-        IProtocolFeesCollector protocolFeesCollector = vault.getProtocolFeesCollector();
+        IProtocolFeesCollector protocolFeesCollector = vault
+            .getProtocolFeesCollector();
 
         // Grant protocolFeesCollector.setSwapFeePercentage role
-        bytes32 actionId = Authentication(address(protocolFeesCollector)).getActionId(
-            protocolFeesCollector.setSwapFeePercentage.selector
-        );
+        bytes32 actionId = Authentication(address(protocolFeesCollector))
+            .getActionId(protocolFeesCollector.setSwapFeePercentage.selector);
         authorizer.grantRole(actionId, address(this));
         protocolFeesCollector.setSwapFeePercentage(0.1e18);
 
@@ -430,10 +453,16 @@ contract SpaceTest is Test {
         }
 
         // No additional lp shares extracted until somebody joins or exits
-        assertEq(space.balanceOf(address(protocolFeesCollector)), 1573707624439152);
+        assertEq(
+            space.balanceOf(address(protocolFeesCollector)),
+            1573707624439152
+        );
         jim.exit(space.balanceOf(address(jim)));
 
-        assertEq(space.balanceOf(address(protocolFeesCollector)), 7502641632334072);
+        assertEq(
+            space.balanceOf(address(protocolFeesCollector)),
+            7502641632334072
+        );
 
         // TODO fees don't eat into non-trade invariant growth
         // TODO fees are correctly proportioned to the fee set in the vault
@@ -481,8 +510,10 @@ contract SpaceTest is Test {
         ava.join();
         // BPT from Ava's (1 Zero, 1 Target) join
         uint256 bptFromJoin = space.balanceOf(address(ava));
-        uint256 targetInFromJoin = INTIAL_USER_BALANCE - target.balanceOf(address(ava));
-        uint256 zeroInFromJoin = INTIAL_USER_BALANCE - zero.balanceOf(address(ava));
+        uint256 targetInFromJoin = INTIAL_USER_BALANCE -
+            target.balanceOf(address(ava));
+        uint256 zeroInFromJoin = INTIAL_USER_BALANCE -
+            zero.balanceOf(address(ava));
 
         vm.warp(1 days);
         uint256 scale1Week = adapter.scale();
@@ -490,11 +521,23 @@ contract SpaceTest is Test {
 
         // Ava's BPT out will exactly equal her first join
         // Since the Target is worth more, she essentially got fewer BPT for the same amount of Underlying
-        assertClose(bptFromJoin, space.balanceOf(address(ava)) - bptFromJoin, 1e3);
+        assertClose(
+            bptFromJoin,
+            space.balanceOf(address(ava)) - bptFromJoin,
+            1e3
+        );
         // Same amount of Target in (but it's worth more now)
-        assertClose(targetInFromJoin * 2, INTIAL_USER_BALANCE - target.balanceOf(address(ava)), 1e3);
+        assertClose(
+            targetInFromJoin * 2,
+            INTIAL_USER_BALANCE - target.balanceOf(address(ava)),
+            1e3
+        );
         // Same amount of Zero in
-        assertClose(zeroInFromJoin * 2, INTIAL_USER_BALANCE - zero.balanceOf(address(ava)), 1e3);
+        assertClose(
+            zeroInFromJoin * 2,
+            INTIAL_USER_BALANCE - zero.balanceOf(address(ava)),
+            1e3
+        );
 
         // Ava can exit her entire LP position just fine
         ava.exit(space.balanceOf(address(ava)));
@@ -505,7 +548,11 @@ contract SpaceTest is Test {
         // There is some change due to the YS invariant, but it's not much in 1 days time
         // With the rate the Target is increasing in value, its growth should account for most of the change
         // in swap rate
-        assertClose(targetOutForOneZeroInit, targetOutForOneZero1Week.mulDown(scale1Week.divDown(initScale)), 1e15);
+        assertClose(
+            targetOutForOneZeroInit,
+            targetOutForOneZero1Week.mulDown(scale1Week.divDown(initScale)),
+            1e15
+        );
     }
 
     function testDifferentDecimals() public {
@@ -514,10 +561,19 @@ contract SpaceTest is Test {
         MockDividerSpace divider = new MockDividerSpace(8);
         // Set Target to 9 decimals
         MockAdapterSpace adapter = new MockAdapterSpace(9);
-        SpaceFactory spaceFactory = new SpaceFactory(vault, address(divider), ts, g1, g2);
+        SpaceFactory spaceFactory = new SpaceFactory(
+            vault,
+            address(divider),
+            ts,
+            g1,
+            g2
+        );
         Space space = Space(spaceFactory.create(address(adapter), maturity));
 
-        (address _zero, , , , , , , , ) = MockDividerSpace(divider).series(address(adapter), maturity);
+        (address _zero, , , , , , , , ) = MockDividerSpace(divider).series(
+            address(adapter),
+            maturity
+        );
         ERC20Mintable zero = ERC20Mintable(_zero);
         ERC20Mintable _target = ERC20Mintable(adapter.target());
 
@@ -549,7 +605,8 @@ contract SpaceTest is Test {
         sid.swapIn(true, 1e18);
         jim.join(1e18, 1e18);
         // Determine Jim's Target balance in 9 decimals
-        uint256 jimTargetBalance = target.balanceOf(address(jim)) / 10**(18 - _target.decimals());
+        uint256 jimTargetBalance = target.balanceOf(address(jim)) /
+            10**(18 - _target.decimals());
 
         assertClose(_target.balanceOf(address(max)), jimTargetBalance, 1e6);
     }
